socketio-sandbox/requirements.txt
click==6.6
Flask==0.11.1
Flask-SocketIO==2.7.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.8
python-engineio==1.0.3
python-socketio==1.6.0
six==1.10.0
Werkzeug==0.11.11
